Updated the register.php file, login.php file  and function page
Added file "doc" at the top of the "functions.php" file,  register.php file and login.php file.

In the "function.php" file all function name converted from the function name being  separated with "under score" to "camel case"

"doc" has been added to all the  functions.

A new function "getUserByUserName()" has been added to the "function.php" file.

In both register.php file and login.php files the database connections has been updated.

In the "login page" the function "getUserByUserName()" is used to first verify if the user exist in the database and if the user exist in the database the "password_verify()" function is used to verify if the "hashed password" matches the password submitted during the login attempt.