Updated The Rester Page
All the code in the "register.php" file has been updated.

The input field data is now validated.

If there is errors in the form data all the errors are stored in an array and variables.

The array and error variable is used to capture what errors have been triggered.

If the error variable and array has a error value the specific error will be printed to the page using the set error variable.

For image field the "$_FILES" SUPERGLOBAL is used to determine if there is a image to uploade.